feat:新建分支保存QGIS32815源码

+++ src/core/numericformats/qgscoordinatenumericformat.cpp
@@ -14,7 +14,6 @@
  ***************************************************************************/
 
 #include "qgscoordinatenumericformat.h"
-#include "moc_qgscoordinatenumericformat.cpp"
 #include "qgis.h"
 #include "qgscoordinateformatter.h"
 
@@ -23,7 +22,6 @@
 #include <locale>
 #include <iomanip>
 
-///@cond PRIVATE
 struct formatter : std::numpunct<wchar_t>
 {
   formatter( QChar thousands, bool showThousands, QChar decimal )
@@ -39,7 +37,7 @@ struct formatter : std::numpunct<wchar_t>
   wchar_t mDecimal;
   bool mShowThousands = true;
 };
-///@endcond
+
 
 QgsGeographicCoordinateNumericFormat::QgsGeographicCoordinateNumericFormat()
 {
@@ -202,7 +200,7 @@ QString QgsGeographicCoordinateNumericFormat::formatLatitudeAsDegreesMinutesSeco
     wrappedY = wrappedY + 180.0;
   }
 
-  const double precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
+  const int precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
 
   int degreesY = int( std::fabs( wrappedY ) );
   const double floatMinutesY = ( std::fabs( wrappedY ) - degreesY ) * 60.0;
@@ -293,7 +291,7 @@ QString QgsGeographicCoordinateNumericFormat::formatLongitudeAsDegreesMinutesSec
     wrappedX = wrappedX + 360.0;
   }
 
-  const double precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
+  const int precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
 
   int degreesX = int( std::fabs( wrappedX ) );
   const double floatMinutesX = ( std::fabs( wrappedX ) - degreesX ) * 60.0;
@@ -393,7 +391,7 @@ QString QgsGeographicCoordinateNumericFormat::formatLatitudeAsDegreesMinutes( do
   int degreesY = int( std::fabs( wrappedY ) );
   double floatMinutesY = ( std::fabs( wrappedY ) - degreesY ) * 60.0;
 
-  const double precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
+  const int precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
 
   //make sure rounding to specified precision doesn't create minutes >= 60
   if ( std::round( floatMinutesY * precisionMultiplier ) >= 60 * precisionMultiplier )
@@ -464,7 +462,7 @@ QString QgsGeographicCoordinateNumericFormat::formatLongitudeAsDegreesMinutes( d
   int degreesX = int( std::fabs( wrappedX ) );
   double floatMinutesX = ( std::fabs( wrappedX ) - degreesX ) * 60.0;
 
-  const double precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
+  const int precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
 
   //make sure rounding to specified precision doesn't create minutes >= 60
   if ( std::round( floatMinutesX * precisionMultiplier ) >= 60 * precisionMultiplier )
@@ -540,7 +538,7 @@ QString QgsGeographicCoordinateNumericFormat::formatLatitudeAsDegrees( double va
 
   const double absY = std::fabs( wrappedY );
 
-  const double precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
+  const int precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
 
   QString hemisphere;
   QString sign;
@@ -592,7 +590,7 @@ QString QgsGeographicCoordinateNumericFormat::formatLongitudeAsDegrees( double v
 
   const double absX = std::fabs( wrappedX );
 
-  const double precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
+  const int precisionMultiplier = std::pow( 10.0, numberDecimalPlaces() );
 
   QString hemisphere;
   QString sign;
