+++ src/gui/editorwidgets/core/qgseditorconfigwidget.cpp
@@ -14,7 +14,6 @@
  ***************************************************************************/
 
 #include "qgseditorconfigwidget.h"
-#include "moc_qgseditorconfigwidget.cpp"
 #include "qgspropertyoverridebutton.h"
 #include "qgsexpressioncontextutils.h"
 
@@ -44,7 +43,7 @@ QgsExpressionContext QgsEditorConfigWidget::createExpressionContext() const
 void QgsEditorConfigWidget::initializeDataDefinedButton( QgsPropertyOverrideButton *button, QgsWidgetWrapper::Property key )
 {
   button->blockSignals( true );
-  button->init( static_cast< int >( key ), mPropertyCollection, QgsWidgetWrapper::propertyDefinitions(), mLayer );
+  button->init( key, mPropertyCollection, QgsWidgetWrapper::propertyDefinitions(), mLayer );
   connect( button, &QgsPropertyOverrideButton::changed, this, &QgsEditorConfigWidget::updateProperty );
   button->registerExpressionContextGenerator( this );
   button->blockSignals( false );
